Superfluid Density & Optimizations
Superfluid density estimators for QMC and ED

Pulled out statistical analysis functions into a separate file for clarity and added functionality to drop samples during post-processing

Improved error messaging in cpp code while reading probability tables

Added the ability to introduce a distance dependant offset in the directed loop probabilities to improve convergence when Delta has a large magnitude and is negative

Script for plotting SSE results

Improved equilibration process so that the number of loops is such that the average number of vertices visited in 1 MC steps is 2<n>

Added functionality to impose Z_2 symmetry when h=0. This helps convergence, especially for magnetization.